requirements.txt
google-genai
genanki